refactor: code cleanup - dead code, renames, merges, file moves
- Delete InstantMessagingProtocol/old/ (extract ImMessage to types.ts)
- Delete consensusAdapter.ts (75% commented-out, unused)
- Rename mempool_v2.ts → mempool.ts (update 16 imports)
- Merge src/exceptions/ into src/errors/ (update 5 imports)
- Move root scripts (run, node-doctor, reset-node, start_db, install-deps.sh) to scripts/
- Move src/features/zk/iZKP/test.ts to src/tests/test_zk_izkp.ts
- Untrack demosStation.sol (Solidity doesn't belong in TS node src/)
- Add *.sol to .gitignore
