+++ scripts/node-doctor
