+++ scripts/reset-node
