--- src/features/InstantMessagingProtocol/old/instantMessagingProtocol.ts
@@ -1,2 +1,0 @@
-import IMSession from "./types/IMSession"
-
