+++ src/libs/network/endpointHandlers.ts
@@ -23,7 +23,7 @@ import _ from "lodash"
 import log from "src/utilities/logger"
 import { getSharedState } from "src/utilities/sharedState"
 import PeerManager from "src/libs/peer/PeerManager"
-import Mempool from "src/libs/blockchain/mempool_v2"
+import Mempool from "src/libs/blockchain/mempool"
 import handleL2PS from "./routines/transactions/handleL2PS"
 import handleDemosWorkRequest from "./routines/transactions/demosWork/handleDemosWorkRequest"
 import multichainDispatcher from "src/features/multichain/XMDispatcher"
